# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Dennic Project",
        "contact": {},
        "version": "1.7"
    },
    "host": "localhost:9050",
    "paths": {
...
        "/v1/archive": {
            "get": {
                "description": "ListArchive - Api for list archive",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Archive"
                ],
                "summary": "ListArchive",
                "parameters": [
                    {
                        "enum": [
                            "status"
                        ],
                        "type": "string",
                        "description": "searchField",
                        "name": "searchField",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "value",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_booking_service.ArchivesType"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            },
            "put": {
                "description": "UpdateArchive - Api for update archive",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Archive"
                ],
                "summary": "UpdateArchive",
                "parameters": [
                    {
                        "description": "UpdateArchiveReq",
                        "name": "UpdateArchiveReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model_booking_service.UpdateArchiveReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_booking_service.Archive"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            },
            "post": {
                "description": "CreateArchive - Api for crete archive",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Archive"
                ],
                "summary": "CreateArchive",
                "parameters": [
                    {
                        "description": "CreateArchiveReq",
                        "name": "CreateArchiveReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model_booking_service.CreateArchiveReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_booking_service.Archive"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            },
            "delete": {
                "description": "DeleteArchive - Api for delete archive",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Archive"
                ],
                "summary": "DeleteArchive",
                "parameters": [
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StatusRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            }
        },
        "/v1/archive/get": {
            "get": {
                "description": "GetArchive - Api for get archive",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Archive"
                ],
                "summary": "GetArchive",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_booking_service.Archive"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model_booking_service.Archive": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "doctor_availability_id": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "patient_problem": {
                    "type": "string"
                },
                "payment_amount": {
                    "type": "number"
                },
                "payment_type": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model_booking_service.ArchivesType": {
            "type": "object",
            "properties": {
                "archives": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model_booking_service.Archive"
                    }
                },
                "count": {
                    "type": "integer"
                }
            }
        },
...
        "model_booking_service.CreateArchiveReq": {
            "type": "object",
            "properties": {
                "doctor_availability_id": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "patient_problem": {
                    "type": "string"
                },
                "payment_amount": {
                    "type": "number"
                },
                "payment_type": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "model_booking_service.UpdateArchiveReq": {
            "type": "object",
            "properties": {
                "archive_id": {
                    "type": "string"
                },
                "doctor_availability_id": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "patient_problem": {
                    "type": "string"
                },
                "payment_amount": {
                    "type": "number"
                },
                "payment_type": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "model_common.ResponseError": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model_common.StandardErrorModel": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/model_common.ResponseError"
                }
            }
        },
...
        "models.StatusRes": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}